feat: remove modal in send form
